+++ src/common/editor/get-active-document-language-id.ts
@@ -1,0 +1,6 @@
+import { getActiveDocument } from "./get-active-document";
+
+export function getActiveDocumentLanguageId(): string | undefined {
+    const document = getActiveDocument();
+    return document ? document.languageId : undefined;
+}