+++ src/common/editor/get-active-document.ts
@@ -1,0 +1,9 @@
+import * as vscode from 'vscode';
+import { getActiveEditor } from './get-active-editor';
+
+export function getActiveDocument(): vscode.TextDocument | undefined {
+    const editor = getActiveEditor();
+    return editor 
+      ? editor.document 
+      : undefined;
+}