+++ src/common/editor/get-active-editor.ts
@@ -1,0 +1,6 @@
+import * as vscode from 'vscode';
+
+export function getActiveEditor(): vscode.TextEditor | undefined {
+    return vscode.window.activeTextEditor;
+} 
+