+++ src/common/nullable.ts
@@ -7,3 +7,7 @@ export function isNullish(value: unknown): value is Nullish {
   return value === null || value === undefined;
 }
 
+export function isUndefined(value: unknown): value is undefined {
+  return typeof(value) === "undefined";
+}
+