+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Change Log
 
-All notable changes to the "toolbox" extension will be documented in this file.
+All notable changes to the "hallpass toolbox" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+The format is influenced by [Keep a Changelog](http://keepachangelog.com/), 
+and we use [Semantic Versioning](https://semver.org/spec/v2.0.0.html) as we grow.
 
 ## [Unreleased]
 
-- Initial release
+### [Added]
+
+- Surround selected text/elements with an HTML tag.
+
+
+## [0.1.0] - 2025-01-19
+
+- Initial Release